Move authentication routes under /auth scope
