+++ backend/src/routes/mod.rs
@@ -1,2 +1,3 @@
+pub mod auth;
 pub mod health_check;
 pub mod users;
