+++ backend/src/startup.rs
@@ -1,5 +1,5 @@
 use crate::{
-    routes::{health_check::health_check_route, users::users_scope},
+    routes::{auth::auth_scope, health_check::health_check_route, users::users_scope},
     util::configuration::{Configuration, DatabaseConfiguration, Environment},
 };
 use actix_session::{storage::CookieSessionStore, SessionMiddleware};
@@ -72,7 +72,8 @@ async fn run(listener: TcpListener, db_pool: MySqlPool, config: Configuration) -
             .service(
                 scope("/api")
                     .service(health_check_route)
-                    .service(scope("/users").configure(users_scope)),
+                    .service(scope("/users").configure(users_scope))
+                    .service(scope("/auth").configure(auth_scope)),
             )
     })
     .listen(listener)?
